+++ spec/rspec/matchers/configuration_spec.rb
@@ -205,7 +205,7 @@ module RSpec
         # config setting, which makes it hard to get at the original
         # default value. in spec_helper.rb we store the default value
         # in $default_expectation_syntax so we can use it here.
-        expect($default_expectation_syntax).to match_array([:expect, :should])
+        expect($default_expectation_syntax).to contain_exactly(:expect, :should)
       end
     end
   end