+++ Changelog.md
@@ -18,6 +18,10 @@ Enhancements:
   Simply chain them off of any existing matcher to create an expression
   like `expect(alphabet).to start_with("a").and end_with("z")`.
   (Eloy Espinaco)
+* Add `contain_exactly` as a less ambiguous version of `match_array`.
+  Note that it expects the expected array to be splatted as
+  individual args: `expect(array).to contain_exactly(1, 2)` is
+  the same as `expect(array).to match_array([1, 2])`. (Myron Marston)
 
 Breaking Changes for 3.0.0: